Fix Bmv2 related issues
- fix pom issues introduced in ONOS-7251
- fix error-prone detected issue
- fix artifactId naming

+++ drivers/bmv2/src/main/java/org/onosproject/drivers/bmv2/impl/Bmv2PreGroupTranslatorImpl.java
@@ -250,7 +250,7 @@ public final class Bmv2PreGroupTranslatorImpl {
     private static void validatePort(PortNumber portNumber) {
         if (portNumber.toLong() < 0 || portNumber.toLong() >= BMV2_PORT_MAP_SIZE) {
             throw new IllegalStateException(String.format("Port number %d is not a valid BMv2 physical port number." +
-                                                             "Valid port range is [0,255]", portNumber));
+                                                             "Valid port range is [0,255]", portNumber.toLong()));
         }
     }
